+++ packages/xo-server/src/models/token.mjs
@@ -2,4 +2,11 @@ import Collection from '../collection/redis.mjs'
 
 // ===================================================================
 
-export class Tokens extends Collection {}
+export class Tokens extends Collection {
+  _unserialize(token) {
+    if (token.created_at !== undefined) {
+      token.created_at = +token.created_at
+    }
+    token.expiration = +token.expiration
+  }
+}