+++ packages/xo-web/src/common/xo/index.js
@@ -1940,6 +1940,8 @@ export const importDisks = (disks, sr) =>
     )
   )
 
+export const getBlockdevices = host => _call('host.getBlockdevices', { id: resolveId(host) })
+
 import ExportVmModalBody from './export-vm-modal' // eslint-disable-line import/first
 export const exportVm = async vm => {
   const { compression, format } = await confirm({
@@ -3485,6 +3487,10 @@ export const updateXosanPacks = pool =>
     return downloadAndInstallXosanPack(pack, pool, { version: pack.version })
   })
 
+// XOSTOR   --------------------------------------------------------------------
+
+export const createXostorSr = params => _call('xostor.create', params)
+
 // Licenses --------------------------------------------------------------------
 
 export const getLicenses = ({ productType } = {}) => _call('xoa.licenses.getAll', { productType })
@@ -3630,6 +3636,11 @@ export const destroyProxyAppliances = proxies =>
 
 export const upgradeProxyAppliance = (proxy, props) =>
   _call('proxy.upgradeAppliance', { id: resolveId(proxy), ...props })
+
+export const openTunnelOnProxy = async proxy => {
+  const result = await _call('proxy.openSupportTunnel', { id: resolveId(proxy) }).catch(err => err.message)
+  await alert(_('supportTunnel'), <pre>{result}</pre>)
+}
 
 export const getProxyApplianceUpdaterState = id => _call('proxy.getApplianceUpdaterState', { id })