+++ packages/xo-web/src/xo-app/xoa/licenses/index.js
@@ -18,7 +18,7 @@ import { get } from '@xen-orchestra/defined'
 import { getLicenses, selfBindLicense, subscribePlugins, subscribeProxies, subscribeSelfLicenses } from 'xo'
 
 import Proxies from './proxies'
-import Xosan from './xosan'
+import Xostor from './xostor'
 
 // -----------------------------------------------------------------------------
 
@@ -196,7 +196,7 @@ export default class Licenses extends Component {
 
     return getLicenses()
       .then(licenses => {
-        const { proxy, xcpng, xoa, xosan } = groupBy(licenses, license => {
+        const { proxy, xcpng, xoa, xosan, xostor } = groupBy(licenses, license => {
           for (const productType of license.productTypes) {
             if (productType === 'xo') {
               return 'xoa'
@@ -210,6 +210,9 @@ export default class Licenses extends Component {
             if (productType === 'xcpng') {
               return 'xcpng'
             }
+            if (productType === 'xostor') {
+              return 'xostor'
+            }
           }
           return 'other'
         })
@@ -219,6 +222,7 @@ export default class Licenses extends Component {
             xcpng,
             xoa,
             xosan,
+            xostor,
           },
         })
       })
@@ -300,6 +304,21 @@ export default class Licenses extends Component {
         }
       })
 
+      // --- XOSTOR ---
+      forEach(licenses.xostor, license => {
+        // When `expires` is undefined, the license isn't expired
+        if (!(license.expires < now)) {
+          products.push({
+            buyer: license.buyer,
+            expires: license.expires,
+            id: license.id,
+            product: 'XOSTOR',
+            type: 'xostor',
+            srId: license.boundObjectId,
+          })
+        }
+      })
+
       return products
     }
   )
@@ -377,18 +396,8 @@ export default class Licenses extends Component {
         </Row>
         <Row>
           <Col>
-            <h2>
-              XOSAN
-              <a
-                className='btn btn-secondary ml-1'
-                href='https://xen-orchestra.com/#!/xosan-home'
-                target='_blank'
-                rel='noopener noreferrer'
-              >
-                <Icon icon='bug' /> {_('productSupport')}
-              </a>
-            </h2>
-            <Xosan xosanLicenses={this.state.licenses.xosan} updateLicenses={this._updateLicenses} />
+            <h2>{_('xostor')}</h2>
+            <Xostor xostorLicenses={this.state.licenses.xostor} updateLicenses={this._updateLicenses} />
           </Col>
         </Row>
         <Row>